# Optimize API collections