--- a/app/models/manageiq/providers/azure/inventory/collector.rb
+++ b/app/models/manageiq/providers/azure/inventory/collector.rb
@@ -17,8 +17,8 @@ class ManageIQ::Providers::Azure::Inventory::Collector < ManagerRefresh::Invento
 
     @config          = manager.connect
     @subscription_id = @config.subscription_id
-    @thread_limit    = options.parallel_thread_limit.to_i || 0
-    @record_limit    = options.targeted_api_collection_threshold.to_i || 500
+    @thread_limit    = (options.parallel_thread_limit || 0)
+    @record_limit    = (options.targeted_api_collection_threshold || 500).to_i
 
     # Caches for optimizing fetching resources and templates of stacks
     @stacks_not_changed_cache = {}
@@ -94,6 +94,15 @@ class ManageIQ::Providers::Azure::Inventory::Collector < ManagerRefresh::Invento
 
   def account_keys(storage_acct)
     @sas.list_account_keys(storage_acct.name, storage_acct.resource_group)
+  end
+
+  def stacks
+    @stacks_cache ||= collect_inventory(:deployments) { stacks_in_parallel(@tds, 'list') }
+
+    stacks_advanced_caching(@stacks_cache) unless @stacks_advanced_caching_done
+    @stacks_advanced_caching_done = true
+
+    @stacks_cache
   end
 
   def stack_templates
@@ -275,5 +284,16 @@ class ManageIQ::Providers::Azure::Inventory::Collector < ManagerRefresh::Invento
       :resource_status_reason => resource.resource_status_reason,
       :last_updated           => resource.last_updated
     }
+  end
+
+  def stacks_in_parallel(arm_service, method_name)
+    region = @ems.provider_region
+
+    Parallel.map(resource_groups, :in_threads => thread_limit) do |resource_group|
+      arm_service.send(method_name, resource_group.name).select do |resource|
+        location = resource.respond_to?(:location) ? resource.location : resource_group.location
+        location.casecmp(region).zero?
+      end
+    end.flatten
   end
 end