--- a/spec/models/manageiq/providers/azure/cloud_manager/azure_refresher_spec_common.rb
+++ b/spec/models/manageiq/providers/azure/cloud_manager/azure_refresher_spec_common.rb
@@ -35,7 +35,6 @@ module AzureRefresherSpecCommon
     }
   ].freeze
 
-
   GRAPH_REFRESH_ADDITIONAL_SETTINGS = [
     {
       :targeted_api_collection_threshold => 0,