… app/models/manageiq/providers/azure/inventory/collector/network_manager.rb
… app/models/manageiq/providers/azure/inventory/collector/target_collection.rb
… app/models/manageiq/providers/azure/inventory/parser/cloud_manager.rb
… app/models/manageiq/providers/azure/inventory_collection_default/cloud_manager.rb
… app/models/manageiq/providers/azure/refresh_helper_methods.rb
… config/settings.yml
… spec/models/manageiq/providers/azure/cloud_manager/deployments_caching_spec.rb
… spec/models/manageiq/providers/azure/cloud_manager/targeted_refresher_scope_spec.rb
… spec/models/manageiq/providers/azure/cloud_manager/targeted_refresher_spec.rb
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/cloud_network_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/lb_created_by_stack_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/lb_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/lb_vms_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/lb_with_vms_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/lbs_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/multiple_targets_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/network_port_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/orchestration_stack_lb_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/orchestration_stack_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/orchestration_stack_vm_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/powered_off_vm_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/powered_on_vm_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/resource_group_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/security_group_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/template_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_0/vm_with_managed_disk_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_500/cloud_network_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_500/lb_created_by_stack_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_500/lb_refresh.yml
… spec/vcr_cassettes/manageiq/providers/azure/cloud_manager/refresher_targeted/targeted_api_collection_threshold_500/lb_vms_refresh.yml