Optimize fetching of stack templates and resources. We will fetch templates and resources only if stack's timestamp has changed. Then add optimizations of storage entities and network entities API collection optimizations.

Partially fixes https://bugzilla.redhat.com/show_bug.cgi?id=1487602
Depends on:
- [x] https://github.com/ManageIQ/manageiq-providers-azure/pull/219


Numbers:
the speed improvement for full refresh is 2200s vs 120s now (with private images included)
the targeted refresh of one Vm takes 15-20s, before it was around 40s

refresh of all Vms at once (targeted refresh) ```EmsRefresh.refresh(Vm.all.to_a)``` takes 81s for 82 VMs, it does a lot of n+1 requests, but we have it configurable threshold to fetch all instead of by 1