@Ladas 2018-03-12
@bronaghs @djberg96  ready for review