@Fryguy 2018-03-12
Does this introduce parallelization with Parallel gem or was that there already?