@Ladas 2018-03-12
@Fryguy https://github.com/Ladas/manageiq-providers-azure/blob/7c3019e7fce77251597a05fc86db8517aed9f936/app/models/manageiq/providers/azure/inventory/collector.rb#L226

Parallel gem is used inside of the armrest, it wasn;t used in collector before