@simaishi 2018-03-22
Gaprindashvili backport details:

```
$ git log -1
commit d2aa713edde2a8b222520485eebdf300105e02c9
Author: Daniel Berger <djberg96@gmail.com>
Date:   Thu Mar 15 15:00:54 2018 -0600

    Merge pull request #220 from Ladas/optimize_fetching_of_stack_templates_and_resources
    
    Optimize API collections
    (cherry picked from commit 83743abfba80cb06fafffb5a4d216eea4d3d2176)
    
    https://bugzilla.redhat.com/show_bug.cgi?id=1558078
```